+++ src/Tracks/UriInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace duncan3dc\Sonos\Tracks;
+
+/**
+ * An interface for objects that repsent some type of Uri
+ */
+interface UriInterface
+{
+
+    /**
+     * Get the URI for this object.
+     *
+     * @return string
+     */
+    public function getUri();
+}