+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
 * [Network] The Network class is no longer static, it should be instantiated before calling its methods.
 * [Network] The cache handling is now provided by doctrine/cache
 * [Playlists] Creating playlists is now done using the createPlaylist() method on the Network class.
+* [Queues/Playlists] Adding individual tracks is now doing using addTrack(), and addTracks() only supports arrays.
 
 --------